+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the ability to configure the underlying transport in Go. #1003
 - Added additional date time (date, time, duration) types in the generation process. #1017
 - PHP Request Adapter (includes middleware) #1048, #918, #1024, #1025
+- Added support for PHP Json Serialization.
+- Update PHP abstractions to make property keys and values nullable in `SerializationWriter.php`.
 
 ### Changed